cli/pyproject.toml
[build-system]
requires = [
    "setuptools>=63.0.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "itm-cli"
description = "Cartken IT-Management CLI"
readme = "README.md"
authors = [
    {name = "Alexander Malyuk", email = "alexander.malyuk@cartken.com"},
]
requires-python = ">= 3.10"
dynamic = [
    "version",
]
dependencies = [
    "typer~=0.12.5",
    "rich~=13.7.1",
    "ansible~=10.4.0",
    "ansible-core~=2.17.5",
    "ansible-runner~=2.4.0",
    "GitPython~=3.1.43",
    "awxkit~=24.6.1",
]

[project.scripts]
itm = "itm_cli.main:main"

[tool.setuptools.dynamic]
version = {attr = "itm_cli.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = [
    "TD",  # todos
    "I",  # isort
]
ignore = [
    "TD003",  # missing-todo-link. Ignored because we don't require adding tickets to TODOs
]
